+++ azext_iot/tests/digitaltwins/__init__.py
@@ -12,8 +12,8 @@ from azure.cli.testsdk import LiveScenarioTest
 from azext_iot.common.embedded_cli import EmbeddedCLI
 
 
-MOCK_RESOURCE_TAGS = "a=b;c=d"
-MOCK_ENDPOINT_TAGS = "key0=value0;key1=value1;"
+MOCK_RESOURCE_TAGS = "a=b c=d"
+MOCK_DEAD_LETTER_SECRET = 'https://accountname.blob.core.windows.net/containerName?sasToken'
 
 
 def generate_resource_id():
@@ -22,8 +22,8 @@ def generate_resource_id():
 
 class DTLiveScenarioTest(LiveScenarioTest):
     role_map = {
-        "owner": "Azure Digital Twins Owner (Preview)",
-        "reader": "Azure Digital Twins Reader (Preview)",
+        "owner": "Azure Digital Twins Data Owner",
+        "reader": "Azure Digital Twins Data Reader",
     }
 
     def __init__(self, test_scenario):