+++ azext_iot/tests/iothub/configurations/test_iot_config_unit.py
@@ -32,11 +32,12 @@ def sample_config_edge_malformed(set_cwd):
     return result
 
 
-@pytest.fixture(params=["file", "inlineA", "inlineB", "layered", "v1"])
+@pytest.fixture(params=["file", "inlineA", "inlineB", "layered", "v1", "v11"])
 def sample_config_edge(set_cwd, request):
     path = "test_edge_deployment.json"
     layered_path = "test_edge_deployment_layered.json"
     v1_path = "test_edge_deployment_v1.json"
+    v11_path = "test_edge_deployment_v11.json"
 
     payload = None
     if request.param == "inlineA":
@@ -49,6 +50,8 @@ def sample_config_edge(set_cwd, request):
         payload = json.dumps(json.loads(read_file_content(layered_path)))
     elif request.param == "v1":
         payload = json.dumps(json.loads(read_file_content(v1_path)))
+    elif request.param == "v11":
+        payload = json.dumps(json.loads(read_file_content(v11_path)))
 
     return (request.param, payload)
 
@@ -289,7 +292,7 @@ class TestConfigCreate:
         assert body.get("priority") == priority
         assert body.get("labels") == evaluate_literal(labels, dict)
 
-        if sample_config_edge[0] == "inlineB":
+        if sample_config_edge[0] == "inlineB" or sample_config_edge[0] == "v11":
             assert (
                 body["content"]["modulesContent"]
                 == json.loads(sample_config_edge[1])["modulesContent"]
@@ -678,7 +681,7 @@ class TestConfigUpdate:
 
 
 class TestConfigList:
-    @pytest.fixture(params=[10, 0, 20])
+    @pytest.fixture(params=[10, 0, 1000])
     def service_client(self, mocked_response, fixture_ghcs, request):
         result = []
         size = request.param
@@ -711,7 +714,7 @@ class TestConfigList:
         mocked_response.expected_size = size
         yield mocked_response
 
-    @pytest.mark.parametrize("top", [1, 100])
+    @pytest.mark.parametrize("top", [1, 100, 1000, None])
     def test_config_list(self, fixture_cmd, service_client, top):
         result = subject.iot_hub_configuration_list(
             cmd=fixture_cmd, hub_name=mock_target["entity"], top=top
@@ -722,9 +725,9 @@ class TestConfigList:
         assert len(result) == top or len(result) == service_client.expected_size * 2
 
         list_request = service_client.calls[0].request
-        assert "top={}".format(top) in list_request.url
+        assert "top=" not in list_request.url
 
-    @pytest.mark.parametrize("top", [1, 10])
+    @pytest.mark.parametrize("top", [1, 100, 1000, None])
     def test_deployment_list(self, fixture_cmd, service_client, top):
         result = subject.iot_edge_deployment_list(
             cmd=fixture_cmd, hub_name=mock_target["entity"], top=top
@@ -734,7 +737,7 @@ class TestConfigList:
         assert len(result) == top or len(result) == service_client.expected_size
 
         list_request = service_client.calls[0].request
-        assert "top={}".format(top) in list_request.url
+        assert "top=" not in list_request.url
 
     @pytest.mark.parametrize("top", [-1, 0, 101])
     def test_config_list_invalid_args(self, fixture_cmd, top):
@@ -792,7 +795,7 @@ class TestConfigApply:
             in url
         )
 
-        if sample_config_edge[0] == "inlineB":
+        if sample_config_edge[0] == "inlineB" or sample_config_edge[0] == "v11":
             assert (
                 body["modulesContent"]
                 == json.loads(sample_config_edge[1])["modulesContent"]