+++ azext_iot/tests/test_iot_ext_int.py
@@ -493,10 +493,10 @@ class TestIoTHubDevices(IoTLiveScenarioTest):
         self.cmd(
             '''iot hub device-identity update -d {} -n {} -g {} --primary-key=""
                     --secondary-key=""'''.format(
-                edge_device_ids[1], LIVE_HUB, LIVE_RG
+                device_ids[4], LIVE_HUB, LIVE_RG
             ),
             checks=[
-                self.check("deviceId", edge_device_ids[1]),
+                self.check("deviceId", device_ids[4]),
                 self.check("status", "enabled"),
                 self.exists("authentication.symmetricKey.primaryKey"),
                 self.exists("authentication.symmetricKey.secondaryKey"),
@@ -507,15 +507,43 @@ class TestIoTHubDevices(IoTLiveScenarioTest):
         self.cmd(
             '''iot hub device-identity update -d {} --login {} --set authentication.symmetricKey.primaryKey=""
                  authentication.symmetricKey.secondaryKey=""'''.format(
-                edge_device_ids[1], self.connection_string
+                device_ids[4], self.connection_string
             ),
             checks=[
-                self.check("deviceId", edge_device_ids[1]),
+                self.check("deviceId", device_ids[4]),
                 self.check("status", "enabled"),
                 self.exists("authentication.symmetricKey.primaryKey"),
                 self.exists("authentication.symmetricKey.secondaryKey"),
             ],
         )
+
+        # Test 'az iot hub device regenerate-key'
+        device = self.cmd(
+            '''iot hub device-identity regenerate-key -d {} -n {} -g {} --kt primary
+                    '''.format(
+                edge_device_ids[1], LIVE_HUB, LIVE_RG
+            ),
+            checks=[
+                self.check("deviceId", edge_device_ids[1])
+            ]
+        ).get_output_in_json()
+
+        # Test swap keys 'az iot hub device regenerate-key'
+        self.cmd(
+            '''iot hub device-identity regenerate-key -d {} -n {} -g {} --kt swap
+                    '''.format(
+                edge_device_ids[1], LIVE_HUB, LIVE_RG
+            ),
+            checks=[
+                self.check("authentication.symmetricKey.primaryKey", device['authentication']['symmetricKey']['secondaryKey']),
+                self.check("authentication.symmetricKey.secondaryKey", device['authentication']['symmetricKey']['primaryKey'])
+            ],
+        )
+
+        # Test 'az iot hub device regenerate-key' with non sas authentication
+        self.cmd("iot hub device-identity regenerate-key -d {} -n {} -g {} --kt secondary"
+                 .format(device_ids[0], LIVE_HUB, LIVE_RG),
+                 expect_failure=True)
 
         sym_conn_str_pattern = r"^HostName={}\.azure-devices\.net;DeviceId={};SharedAccessKey=".format(
             LIVE_HUB, edge_device_ids[0]